Move Split and Topology to own file as they're shared among tree spaces
